move editor outputs to a separate component, minor re-org
