+++ frontend/src/components/EditorView.jsx
@@ -1,15 +1,16 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import { actions } from '../store';
 
 import Editor from './Editor';
+import EditorOutputs from './EditorOutputs';
 
 const EditorPage = ({
   currentQ,
   testResults,
   getTestResults
 }) => {
-  const funcString = `function ${currentQ.question.functionName}(${currentQ.question.params}) {\n  \n}`;
+  const funcString = `function ${currentQ.question.functionName}(${currentQ.question.params}) {\n  // Your code here...\n}`;
   const [editorValue, setEditorValue] = useState(
     funcString
   );
@@ -17,59 +18,27 @@ const EditorPage = ({
     setEditorValue(ev.doc.getValue());
   };
 
-  const countWrongResults = results => {
-    return results.filter(result => !!result.wrong).length;
-  };
-
   const testCode = () => {
     getTestResults(currentQ.questionId, editorValue);
   };
 
+  const submitCode = () => {
+    return null;
+  };
+
   return (
-    <div>
+    <div id="editor-view">
       <Editor
         value={editorValue}
         onChange={handleEditorChange}
       />
-      <div id="console-logs">
-        {testResults.length &&
-        testResults[0].consoles.length ? (
-          <div>
-            <h4>Log:</h4>
-            <ul>
-              {testResults[0].consoles.map(c => (
-                <li>{c}</li>
-              ))}
-            </ul>
-          </div>
-        ) : (
-          ''
-        )}
-      </div>
-      <div id="test-results">
-        {testResults
-          ? testResults.map(output => {
-              if (output.wrong)
-                return (
-                  <p className="wrong">{output.wrong}</p>
-                );
-              return (
-                <p className="right">{output.result}</p>
-              );
-            })
-          : ''}
-        {testResults.length &&
-        !countWrongResults(testResults) ? (
-          <h4>All tests pass!</h4>
-        ) : (
-          ''
-        )}
-      </div>
-
       <button type="button" onClick={testCode}>
         Test
       </button>
-      <button type="button">Submit</button>
+      <button type="button" onClick={submitCode}>
+        Submit
+      </button>
+      <EditorOutputs testResults={testResults} />
     </div>
   );
 };
